Refactor MendelException
