--- src/main/java/MendelAction.java
@@ -1,3 +1,0 @@
-public abstract class MendelAction {
-    public abstract void speak();
-}
