--- src/main/java/MendelException.java
@@ -1,10 +1,0 @@
-public class MendelException extends RuntimeException {
-    public MendelException(String message) {
-
-        super(message);
-    }
-    @Override
-    public String toString() {
-        return super.getMessage();
-    }
-}
