+++ src/main/java/mendel/mendelexception/MendelException.java
@@ -1,0 +1,12 @@
+package mendel.mendelexception;
+
+public class MendelException extends RuntimeException {
+  public MendelException(String message) {
+
+    super(message);
+  }
+  @Override
+  public String toString() {
+    return super.getMessage();
+  }
+}
