+++ src/main/java/mendel/metacognition/Leave.java
@@ -1,0 +1,4 @@
+package mendel.metacognition;
+
+public class Leave {
+}
