+++ src/main/java/mendel/metacognition/MendelAction.java
@@ -1,0 +1,4 @@
+package mendel.metacognition;
+
+public class MendelAction {
+}
