+++ src/main/java/mendel/metacognition/TaskStorage.java
@@ -1,0 +1,4 @@
+package mendel.metacognition;
+
+public class TaskStorage {
+}
