+++ src/main/java/mendel/metacognition/Welcome.java
@@ -1,0 +1,4 @@
+package mendel.metacognition;
+
+public class Welcome {
+}
